Add a bunch of packages I'll probably need
I'm about to go offline so I thought I'd grab a bunch of things that the
existing client code relies on before doing so.

I've gone with `temporal-polyfill` over `moment` because I only needed
durations and Moment is such a very big hammer to use for durations.

I'd like to believe that some day this'll be in the stdlib but who
knows, JS seems determined to be horrible.